STEP-4: initialization of microservices (adding build.gradle files)
