+++ build.gradle
@@ -1,19 +1,20 @@
 plugins {
-    id("java")
+    id 'java'
 }
 
-group = "org.example"
-version = "1.0-SNAPSHOT"
+group 'org.example'
+version '1.0-SNAPSHOT'
 
 repositories {
     mavenCentral()
 }
 
 dependencies {
-    testImplementation(platform("org.junit:junit-bom:5.9.1"))
-    testImplementation("org.junit.jupiter:junit-jupiter")
+    testImplementation 'junit:junit:4.13.2'
 }
+
 
 tasks.test {
     useJUnitPlatform()
 }
+
