+++ compensation-device-service/settings.gradle
@@ -1,0 +1,2 @@
+rootProject.name = 'compensation-device-service'
+
