--- compensation-device-service/src/main/java/CompensationDeviceApplication.java
@@ -1,2 +1,0 @@
-public class CompensationDeviceApplication {
-}
