--- config-service/src/main/java/ConfigApplication.java
@@ -1,2 +1,0 @@
-public class ConfigApplication {
-}
