+++ full-information-service/build.gradle
@@ -1,17 +1,46 @@
 plugins {
-    id("java")
+    id 'java'
+    id 'io.spring.dependency-management' version '1.0.8.RELEASE'
+    id "io.freefair.lombok" version "5.1.0"
 }
 
-group = "org.example"
 version = "1.0-SNAPSHOT"
 
 repositories {
     mavenCentral()
 }
 
+jar {
+    manifest {
+        attributes(
+                'Main-Class': 'com.fullinformationservice.FullInformationApplication'
+        )
+    }
+}
+
+ext {
+    set('springCloudVersion', "Hoxton.SR1")
+}
+
+dependencyManagement {
+    imports {
+        mavenBom "org.springframework.cloud:spring-cloud-dependencies:${springCloudVersion}"
+    }
+}
+
 dependencies {
-    testImplementation(platform("org.junit:junit-bom:5.9.1"))
-    testImplementation("org.junit.jupiter:junit-jupiter")
+    implementation 'org.springframework.boot:spring-boot:2.3.2.RELEASE'
+    implementation "org.springframework.boot:spring-boot-starter-web:2.3.2.RELEASE"
+    implementation "org.springframework.boot:spring-boot-starter-actuator:2.3.2.RELEASE"
+    implementation "org.springframework.boot:spring-boot-starter-data-jpa:2.3.2.RELEASE"
+    implementation "org.springframework.cloud:spring-cloud-starter-config"
+    implementation "org.springframework.cloud:spring-cloud-starter-openfeign"
+    implementation "org.springframework.cloud:spring-cloud-starter-netflix-eureka-client"
+    implementation "org.springframework.cloud:spring-cloud-starter-sleuth"
+    implementation "org.springframework.cloud:spring-cloud-starter-netflix-hystrix"
+    implementation "org.postgresql:postgresql:42.5.1"
+
+    testImplementation 'org.springframework.boot:spring-boot-starter-test:2.3.2.RELEASE'
 }
 
 tasks.test {
