+++ full-information-service/settings.gradle
@@ -1,0 +1,2 @@
+rootProject.name = 'full-information-service'
+
