--- full-information-service/src/main/java/FullInformationApplication.java
@@ -1,2 +1,0 @@
-public class FullInformationApplication {
-}
