+++ full-information-service/src/main/java/com/fullinformationservice/FullInformationApplication.java
@@ -1,0 +1,13 @@
+package com.fullinformationservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+@SpringBootApplication
+@EnableDiscoveryClient
+public class FullInformationApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(FullInformationApplication.class, args);
+    }
+}
