+++ gateway/build.gradle
@@ -1,19 +1,45 @@
 plugins {
-    id("java")
+    id 'java'
+    id 'io.spring.dependency-management' version '1.0.8.RELEASE'
 }
 
-group = "org.example"
-version = "1.0-SNAPSHOT"
+version '1.0-SNAPSHOT'
+sourceCompatibility = '1.8'
 
 repositories {
     mavenCentral()
 }
 
+jar {
+    manifest {
+        attributes(
+                'Main-Class': 'com.gateway.GatewayApplication'
+        )
+    }
+}
+
+ext {
+    set('springCloudVersion', "Hoxton.SR1")
+}
+
+dependencyManagement {
+    imports {
+        mavenBom "org.springframework.cloud:spring-cloud-dependencies:${springCloudVersion}"
+    }
+}
+
 dependencies {
-    testImplementation(platform("org.junit:junit-bom:5.9.1"))
-    testImplementation("org.junit.jupiter:junit-jupiter")
+    implementation 'org.springframework.boot:spring-boot:2.3.2.RELEASE'
+    implementation "org.springframework.cloud:spring-cloud-starter-netflix-zuul"
+    implementation "org.springframework.cloud:spring-cloud-starter-config"
+    implementation "org.springframework.cloud:spring-cloud-starter"
+    implementation "org.springframework.cloud:spring-cloud-starter-netflix-eureka-client"
+    implementation "org.springframework.cloud:spring-cloud-starter-sleuth"
+
+    testImplementation 'org.springframework.boot:spring-boot-starter-test:2.3.2.RELEASE'
 }
 
 tasks.test {
     useJUnitPlatform()
 }
+
