--- gateway/src/main/java/GatewayApplication.java
@@ -1,2 +1,0 @@
-public class GatewayApplication {
-}
