+++ high-volt-cables-service/settings.gradle
@@ -1,0 +1,2 @@
+rootProject.name = 'high-volt-cables-service'
+
