--- high-volt-cables-service/src/main/java/HighVoltCablesApplication.java
@@ -1,2 +1,0 @@
-public class HighVoltCablesApplication {
-}
