--- lighting-information-service/src/main/java/LightInformationApplication.java
@@ -1,2 +1,0 @@
-public class LightInformationApplication {
-}
