+++ lighting-information-service/src/main/java/com/lightinginformationservice/LightInformationApplication.java
@@ -1,0 +1,13 @@
+package com.lightinginformationservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+@SpringBootApplication
+@EnableDiscoveryClient
+public class LightInformationApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(LightInformationApplication.class, args);
+    }
+}
