--- power-transformers-service/src/main/java/PowerTransformerApplication.java
@@ -1,2 +1,0 @@
-public class PowerTransformerApplication {
-}
