+++ power-transformers-service/src/main/java/com/powertransformerservice/PowerTransformerApplication.java
@@ -1,0 +1,13 @@
+package com.powertransformerservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+@SpringBootApplication
+@EnableDiscoveryClient
+public class PowerTransformerApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(PowerTransformerApplication.class, args);
+    }
+}
