+++ protective-equipment-service/lombok.config
@@ -1,0 +1,2 @@
+# This file is generated by the 'io.freefair.lombok' Gradle plugin
+config.stopBubbling = true
