+++ protective-equipment-service/settings.gradle
@@ -1,0 +1,2 @@
+rootProject.name = 'protective-equipment-service'
+
