--- protective-equipment-service/src/main/java/ProtectiveEquipmentApplication.java
@@ -1,2 +1,0 @@
-public class ProtectiveEquipmentApplication {
-}
