+++ protective-equipment-service/src/main/java/com/protectiveequipmentservice/ProtectiveEquipmentApplication.java
@@ -1,0 +1,13 @@
+package com.protectiveequipmentservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+@SpringBootApplication
+@EnableDiscoveryClient
+public class ProtectiveEquipmentApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(ProtectiveEquipmentApplication.class, args);
+    }
+}
