--- registry/src/main/java/RegistryApplication.java
@@ -1,2 +1,0 @@
-public class RegistryApplication {
-}
