+++ registry/src/main/java/com/registry/RegistryApplication.java
@@ -1,0 +1,13 @@
+package com.registry;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
+
+@SpringBootApplication
+@EnableEurekaServer
+public class RegistryApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(RegistryApplication.class, args);
+    }
+}
