--- start-information-service/src/main/java/StartInformationApplication.java
@@ -1,2 +1,0 @@
-public class StartInformationApplication {
-}
