+++ start-information-service/src/main/java/com/startinformationservice/StartInformationApplication.java
@@ -1,0 +1,13 @@
+package com.startinformationservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+@SpringBootApplication
+@EnableDiscoveryClient
+public class StartInformationApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(StartInformationApplication.class, args);
+    }
+}
